requirements.txt
pymupdf
llama-cpp-python
numpy
matplotlib
tqdm
